+++ src/renderer/components/Expense/Calendar.vue
@@ -1,0 +1,4 @@
+<template>
+  <div>Calendar View</div>
+</template>
+<script></script>